bug: 로그인 성공한 사용자가 onboarding에서 refetch 전에 페이지 이동하는 문제 해결
기존 onboarding 페이지에서 사용자 정보 refetch하는 로직이 정상동작하지 않아, nickname 페이지에서 사용자 정보를 확인하는 로직으로 변경했습니다. 이에 따라 불필요한 코드 또한 정리했습니다.
