+++ src/hooks/reactQuery/auth/useGetMemberData.ts
@@ -3,14 +3,9 @@ import { useQuery } from '@tanstack/react-query';
 import { api } from '@/apis';
 import type { MemberProps } from '@/features/member/types';
 
-interface UseGetMemberDataProps {
-  enabled?: boolean;
-}
-
-export const useGetMemberData = ({ enabled = true }: UseGetMemberDataProps = {}) => {
+export const useGetMemberData = () => {
   return useQuery<MemberProps>({
     queryKey: ['memberData'],
     queryFn: () => api.get<MemberProps>('/my'),
-    enabled,
   });
 };
